+++ apps/test_planar_2dof.cpp
@@ -43,7 +43,7 @@ int main(int argc, char **argv)
 	}
 
 	std::string project_path(__FILE__);
-    for (int i = 0; i < 2; i++) {
+    for (size_t i = 0; i < 2; i++) {
         project_path = project_path.substr(0, project_path.find_last_of("/\\"));
 	}
