+++ apps/test_rbtconnect.cpp
@@ -18,7 +18,7 @@ int main(int argc, char **argv)
 	// std::string scenario_file_path = "/data/xarm6/scenario2/scenario2.yaml";
 	// std::string scenario_file_path = "/data/xarm6/scenario3/scenario3.yaml";
 
-	int max_num_tests = 10;
+	size_t max_num_tests = 10;
 
 	// -------------------------------------------------------------------------------------- //
 	
@@ -29,9 +29,9 @@ int main(int argc, char **argv)
 	const std::string project_path = getProjectPath();
 	ConfigurationReader::initConfiguration(project_path);
     YAML::Node node = YAML::LoadFile(project_path + scenario_file_path);
-	int num_random_obstacles = node["random_obstacles"]["num"].as<int>();
+	size_t num_random_obstacles = node["random_obstacles"]["num"].as<size_t>();
 	Eigen::Vector3f obs_dim;
-	for (int i = 0; i < 3; i++)
+	for (size_t i = 0; i < 3; i++)
 		obs_dim(i) = node["random_obstacles"]["dim"][i].as<float>();
 		
 	scenario::Scenario scenario(scenario_file_path, project_path);
@@ -42,7 +42,7 @@ int main(int argc, char **argv)
 	std::unique_ptr<planning::AbstractPlanner> planner;
 
 	bool result = false;
-	int num_obs = env->getNumObjects();
+	size_t num_obs = env->getNumObjects();
 	while (!result && num_random_obstacles > 0)
 	{
 		env->removeObjects(num_obs);
@@ -59,8 +59,8 @@ int main(int argc, char **argv)
 	LOG(INFO) << "Start: " << q_start;
 	LOG(INFO) << "Goal: " << q_goal;
 
-	int num_test = 0;
-	int num_success = 0;
+	size_t num_test = 0;
+	size_t num_success = 0;
 	std::vector<float> planning_times;
 	while (num_test++ < max_num_tests)
 	{
@@ -81,7 +81,7 @@ int main(int argc, char **argv)
 				planning_times.emplace_back(planner->getPlannerInfo()->getPlanningTime());
 				// LOG(INFO) << "Found path: ";
 				// std::vector<std::shared_ptr<base::State>> path = planner->getPath();
-				// for (int i = 0; i < path.size(); i++)
+				// for (size_t i = 0; i < path.size(); i++)
 				// 	std::cout << i << ": " << path.at(i)->getCoord().transpose() << std::endl;
 			}
