+++ include/scenario/Scenario.h
@@ -30,7 +30,7 @@ namespace scenario
         inline std::shared_ptr<base::State> getStart() const { return q_start; }
         inline std::shared_ptr<base::State> getGoal() const { return q_goal; }
         inline base::StateSpaceType getStateSpaceType() const { return state_space_type; }
-        inline int getNumDimensions() const { return num_dimensions; }
+        inline size_t getNumDimensions() const { return num_dimensions; }
 
         inline void setStart(const std::shared_ptr<base::State> q_start_) { q_start = q_start_; }
         inline void setGoal(const std::shared_ptr<base::State> q_goal_) { q_goal = q_goal_; }
@@ -42,7 +42,7 @@ namespace scenario
         std::shared_ptr<base::State> q_start;
         std::shared_ptr<base::State> q_goal;
         base::StateSpaceType state_space_type;
-        int num_dimensions;         // Number of dimensions of the state-space (also 'num_DOFs' of the used robot)
+        size_t num_dimensions;         // Number of dimensions of the state-space (also 'num_DOFs' of the used robot)
 	};
 }
 #endif //RPMPL_SCENARIO_H
